Change suboption separator to ':' to prepare for JPF rev 82d0030dcd42
JPF rev 82d0030dcd42 introduced changes in how arguments to the app to
be analized are processed.  Unfortunately, these changes have the side
effect that app arguments are split at ',' characters, hence it breaks
cJUnit, because we use ',' to separate the test name and expected
exception suboptions of the '--test' argument.

To work this around we change the suboption separator character from
',' to ':'.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>