Update JPF jars to rev dbda1e61ec94 and related fix
JPF rev dbda1e61ec94 changed the default properties for reporting
property violations in a way that JPF does not include the trace
during which the property was violated anymore.  I.e. the change did:

  -jpf.report.console.property_violation=error,trace,snapshot
  +jpf.report.console.property_violation=error,snapshot

However, having that trace is good, so override this change by setting
the old value in JPFInvoker.setConfigArgs().

Signed-off-by: SZEDER Gábor <szeder@fzi.de>